Remove interactions with Xlib and python3-xlib dependency
The state of Xlib/XCB Python bindings is dismal.
python-xlib/python3-xlib are simply broken with recent RandR versions.
xpyb, which is the default XCB Python binding, is very out of date and
doesn't have a proper Python 3 support (upstream state is "untested").
xcffib doesn't provide as wide coverage as xpyb, but supports Python3
and can be considered almost usable, but there are still some issues
and it requires Haskell toolchain in order to be built.

Thus drop all the code that interacts with Xlib.

This breaks backwards compatibility, as you now must initialize I3SOCK
environment variable properly if you want i3ipc to be able to figure out
the path to the i3 socket file automatically. Compatibility is restored
in the next commit.