aio: Use bytearray instead of bytes in Connection._message
By making the message buffer a `bytearray` instead of `bytes` we can
append to the buffer without having to create a new instance for every
iteration. To avoid another copy we return this `bytearray` instead of
casting to `bytes`.

For consistency all return paths now return `bytearray` instead of
`bytes`.  This swap should not be an issue since all callers pass this
return value to `json.loads` which supports both `bytes` and
`bytearray`.